refactor: ponytail Batch 2 — merge coverage test files into main tests
Consolidate 8 *_coverage_test.go files into package-native *_test.go
entry points. Drop trivial TestDiagLogger_Nop and nil-receiver-only test.

Part of #140. Closes #127.
